changed folder structure to match react/redux best practices, updated CSS names to follow BEM convention, and rendered the quill component!
