+++ app/assets/stylesheets/note_editor.scss
@@ -1,0 +1,8 @@
+.noteEditor {
+    width: 80vw;
+    height: 100vh;
+}
+
+.noteEditor__quill {
+    height: 100%;
+}
