+++ app/views/layouts/application.html.erb
@@ -4,9 +4,9 @@
     <title>Noteworthy</title>
     <%= csrf_meta_tags %>
     <%= csp_meta_tag %>
-
     <%= stylesheet_link_tag    'application', media: 'all' %>
     <%= javascript_include_tag 'application' %>
+    <link rel="stylesheet" href="//cdn.quilljs.com/1.2.6/quill.snow.css">
   </head>
 
   <body>
