+++ config/application.rb
@@ -15,5 +15,8 @@ module Noteworthy
     # Application configuration can go into files in config/initializers
     # -- all .rb files in that directory are automatically loaded after loading
     # the framework and any gems in your application.
+
+    # joining the fonts folder to assets
+    config.assets.paths << Rails.root.join("app", "assets", "fonts")
   end
 end
