--- frontend/components/note_app/note_app_container.jsx
@@ -1,14 +1,0 @@
-import React from "react";
-// import AuthenticationContainer from "../authentication/authentication_container";
-import { Link } from 'react-router-dom';
-// import logo from '../../../app/assets/images/evernote-logo.png';
-
-const noteAppContainer = () => {
-    return (
-        <div className="note-app-container">
-            <h1>the app will go here</h1>
-        </div>
-    );
-}
-
-export default noteAppContainer;
