--- frontend/components/session_form/login_form_container.jsx
@@ -1,21 +1,0 @@
-import {connect} from 'react-redux';
-import React from 'react';
-import {Link} from 'react-router-dom';
-import {login} from '../../actions/session_actions';
-import SessionForm from './session_form';
-
-const mapStateToProps = ({errors}) => {
-    return {
-        errors: errors.session,
-        formType: 'login',
-        navLink: <Link to="/signup">SignUp</Link>
-    }
-};
-
-const mapDispatchToProps = (dispatch) => {
-    return {
-        processForm: (user) => dispatch(login(user)),
-    }
-};
-
-export default connect(mapStateToProps, mapDispatchToProps)(SessionForm);
