--- frontend/components/session_form/signup_form_container.jsx
@@ -1,22 +1,0 @@
-import { connect } from 'react-redux';
-import React from 'react';
-import { Link } from 'react-router-dom';
-import { signup } from '../../actions/session_actions';
-import SessionForm from './session_form';
-
-const mapStateToProps = ({ errors }) => {
-    return {
-        // clarify .session
-        errors: errors.session,
-        formType: 'signup',
-        navLink: <Link to="/login">login</Link>
-    }
-}
-
-const mapDispatchToProps = (dispatch) => {
-    return {
-        processForm: (user) => dispatch(signup(user)),
-    }
-}
-
-export default connect(mapStateToProps, mapDispatchToProps)(SessionForm)
