--- frontend/components/splash_navbar/splash_navbar.jsx
@@ -1,11 +1,0 @@
-
-import React from "react";
-const SplashNavbar = () => {
-    return (
-        <>
-            
-        </>
-    );
-}
-
-export default SplashNavbar;
