--- frontend/index.jsx
@@ -1,28 +1,0 @@
-import React from 'react';
-import ReactDOM from 'react-dom';
-import configureStore from './store/store';
-import Root from './components/root';
-// import 'normalize.css';
-
-document.addEventListener('DOMContentLoaded', () => {
-  let store;
-  
-  // checks if user is logged in already for configuration of store
-  if(window.currentUser) {
-    const preloadedState = {
-      entities: {
-        users: { [window.currentUser.id]: window.currentUser }
-      },
-      session: { id: window.currentUser.id }
-    };
-    store = configureStore(preloadedState);
-    delete window.currentUser;
-  } else {
-    store = configureStore();
-  }
-
-  window.getState = store.getState
-
-  const root = document.getElementById('root');
-  ReactDOM.render(<Root store={store} />, root);
-});
