--- frontend/reducers/entities_reducer.js
@@ -1,8 +1,0 @@
-import { combineReducers } from 'redux';
-import usersReducer from './users_reducer'
-
-const entitiesReducer = combineReducers({
-    users: usersReducer,
-})
-
-export default entitiesReducer;
