--- frontend/reducers/errors_reducer.js
@@ -1,8 +1,0 @@
-import { combineReducers } from 'redux';
-import sessionErrorsReducer from "./sessions_errors_reducer";
-
-const errorsReducer = combineReducers({
-    session: sessionErrorsReducer,
-});
-
-export default errorsReducer;
