--- frontend/reducers/root_reducer.js
@@ -1,12 +1,0 @@
-import { combineReducers } from 'redux';
-import entitiesReducer from './entities_reducer';
-import sessionReducer from './session_reducer';
-import errorsReducer from './errors_reducer';
-
-const rootReducer = combineReducers({
-    entities: entitiesReducer,
-    session: sessionReducer,
-    errors: errorsReducer,
-})
-
-export default rootReducer;
