+++ package.json
@@ -13,6 +13,7 @@
     "babel-loader": "^8.0.4",
     "react": "^16.2.0",
     "react-dom": "^16.2.0",
+    "react-quill": "^1.3.5",
     "react-redux": "^5.0.6",
     "react-router-dom": "^4.2.2",
     "redux": "^3.7.2",
