+++ src/routes/root.jsx
@@ -1,0 +1,14 @@
+import React from "react";
+import { Provider } from "react-redux";
+import { HashRouter } from "react-router-dom";
+import App from "./index";
+
+const Root = ({ store }) => (
+    <Provider store={store}>
+        <HashRouter>
+            <App/>
+        </HashRouter>
+    </Provider>
+);
+
+export default Root;
