+++ src/state/reducers/entities.js
@@ -1,0 +1,6 @@
+import { combineReducers } from 'redux';
+import users from './users'
+
+export default combineReducers({
+    users
+})
