+++ src/state/reducers/errors.js
@@ -1,0 +1,6 @@
+import { combineReducers } from 'redux';
+import session from "./sessionErrors";
+
+export default combineReducers({
+    session
+});
