+++ src/state/reducers/root.js
@@ -1,0 +1,12 @@
+import { combineReducers } from 'redux';
+import entities from './entities';
+import session from './session';
+import errors from './errors';
+
+export default combineReducers({
+    entities,
+    session,
+    errors,
+})
+
+// export default rootReducer;
