+++ src/state/reducers/users.js
@@ -1,0 +1,12 @@
+import { RECEIVE_CURRENT_USER } from "../actions/session";
+
+// if there is a signup/login, set the slice of state equal to the current user
+export default (state = {}, action) => {
+    Object.freeze(state);
+    switch(action.type) {
+        case RECEIVE_CURRENT_USER:
+            return Object.assign({}, state, { [action.currentUser.id]: action.currentUser });
+        default:
+            return state;
+    }
+};
