+++ src/state/util/route.jsx
@@ -1,0 +1,26 @@
+import {withRouter, Route, Redirect} from 'react-router-dom';
+import {connect} from 'react-redux';
+import React from 'react';
+
+// custom routes
+// auth route to protect routes between logged in/not logged in users
+const Guest = ({ component: Component, path, loggedIn, exact }) => {
+    return (<Route
+        path={path}
+        exact={exact}
+        render={props =>
+            !loggedIn ? <Component {...props} /> : <Redirect to="/app" />
+        }
+    />);
+};
+
+// checks if session id is present, meaning user is logged in, and passes to Auth
+const mapStateToProps = (state) => {
+    return { loggedIn: Boolean(state.session.id) };
+};
+  
+export const GuestRoute = withRouter(
+    connect(
+        mapStateToProps
+    )(Guest)
+);
