+++ webpack.config.js
@@ -2,7 +2,7 @@ const path = require('path');
 
 module.exports = {
   context: __dirname,
-  entry: './frontend/index.jsx',
+  entry: './src/index.jsx',
   output: {
     path: path.resolve(__dirname, 'app', 'assets', 'javascripts'),
     filename: 'bundle.js'
@@ -19,12 +19,6 @@ module.exports = {
           loader: 'babel-loader',
           query: {
             presets: ['@babel/env', '@babel/react'],
-            // trying to fix this generatorRuntime error
-            // presets: [[
-            //   "@babel/preset-env", {
-            //     useBuiltIns: "entry"
-            //   }],
-            //   "@babel/preset-react"]
           }
         },
       }
